correct dtype when adding nan-read

+++ haplokit/io/bam.py
@@ -169,8 +169,8 @@ def add_nan_read_if_empty(locus, symbols, quals):
     assert np.size(symbols) == np.size(quals)
     n_reads = symbols.shape[0]
     if n_reads == 0:
-        symbols = np.array([['-'] * len(locus.variants) ])
-        quals = np.zeros((1, len(locus.variants)), dtype=np.int16)
+        symbols = np.array([['-'] * len(locus.variants) ], dtype=symbols.dtype)
+        quals = np.zeros((1, len(locus.variants)), dtype=quals.dtype)
     return symbols, quals
